Time: 0 ms (100%), Space: 25.9 MB (7.8%) - LeetHub

+++ 3471-minimum-average-of-smallest-and-largest-elements/3471-minimum-average-of-smallest-and-largest-elements.cpp
@@ -11,7 +11,7 @@ public:
             ans.push_back(avg);
             i++,j--;
         }
-        double min=DBL_MAX;
+        double min=INT_MAX;
         for(int i=0;i<ans.size();i++)
         {
             if(min>ans[i])
